Extract Msf::DBManager::Import::Wapiti
MSP-11124
